+++ CHANGELOG.md
@@ -26,10 +26,12 @@ Bug Fixes
 Deprecated Features
 -------------------
 
+- Deprecated passing Image arguments to kappaKaiserSquires function. (#855)
 
 
 New Features
 ------------
+
 - Add option to use circular weight function in HSM adaptive moments code. (#917)
 - Add VonKarman profile GSObject.